Move Verkle node types to `database::verkle::variants::managed` module
Since we plan to support other trie types in the future, these should
not live in the generic `types` module.
