+++ rust/src/database/mod.rs
@@ -8,6 +8,6 @@
 // On the date above, in accordance with the Business Source License, use of
 // this software will be governed by the GNU Lesser General Public License v3.
 
-mod verkle;
+pub mod verkle;
 
 pub use verkle::{SimpleInMemoryVerkleTrie, VerkleTrieCarmenState};
