+++ rust/src/database/verkle/variants/managed/nodes/sparse_leaf.rs
@@ -1,0 +1,72 @@
+// Copyright (c) 2025 Sonic Operations Ltd
+//
+// Use of this software is governed by the Business Source License included
+// in the LICENSE file and at soniclabs.com/bsl11.
+//
+// Change Date: 2028-4-16
+//
+// On the date above, in accordance with the Business Source License, use of
+// this software will be governed by the GNU Lesser General Public License v3.
+
+use zerocopy::{FromBytes, Immutable, IntoBytes, Unaligned};
+
+use crate::{database::verkle::crypto::Commitment, types::Value};
+
+/// A value of a leaf node in a (file-based) Verkle trie, together with its index.
+// NOTE: Changing the layout of this struct will break backwards compatibility of the
+// serialization format.
+#[derive(
+    Debug, Clone, Copy, Default, PartialEq, Eq, FromBytes, IntoBytes, Immutable, Unaligned,
+)]
+#[repr(C)]
+pub struct ValueWithIndex {
+    /// The index of the value in the leaf node.
+    pub index: u8,
+    /// The value stored in the leaf node.
+    pub value: Value,
+}
+
+/// A sparsely populated leaf node in a (file-based) Verkle trie.
+// NOTE: Changing the layout of this struct will break backwards compatibility of the
+// serialization format.
+#[derive(Debug, Clone, PartialEq, Eq, FromBytes, IntoBytes, Immutable)]
+#[repr(C)]
+pub struct SparseLeafNode<const N: usize> {
+    pub commitment: Commitment,
+    pub stem: [u8; 31],
+    pub values: [ValueWithIndex; N],
+}
+
+impl<const N: usize> Default for SparseLeafNode<N> {
+    fn default() -> Self {
+        let mut values = [ValueWithIndex::default(); N];
+        values.iter_mut().enumerate().for_each(|(i, v)| {
+            v.index = i as u8;
+        });
+
+        SparseLeafNode {
+            commitment: Commitment::default(),
+            stem: [0; 31],
+            values,
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn sparse_leaf_node_default_returns_leaf_node_with_default_values_and_unique_indices() {
+        const N: usize = 2;
+        let node: SparseLeafNode<N> = SparseLeafNode::default();
+
+        assert_eq!(node.commitment, Commitment::default());
+        assert_eq!(node.stem, [0; 31]);
+
+        for (i, value) in node.values.iter().enumerate() {
+            assert_eq!(value.index, i as u8);
+            assert_eq!(value.value, Value::default());
+        }
+    }
+}
