+++ rust/src/types/node_size.rs
@@ -8,7 +8,11 @@
 // On the date above, in accordance with the Business Source License, use of
 // this software will be governed by the GNU Lesser General Public License v3.
 
-pub mod managed;
-mod simple;
+/// A trait to determine the size of a node in a trie.
+pub trait NodeSize {
+    /// Returns the size of the node in bytes.
+    fn node_byte_size(&self) -> usize;
 
-pub use simple::SimpleInMemoryVerkleTrie;
+    /// Returns the minimum size of a non-empty node in bytes.
+    fn min_non_empty_node_size() -> usize;
+}
